# perf: small scalarmul