# Description

This PR implements switching scalar multiplication to windowed method when scalars are small (according to some heuristic).

Fixes #702

[g1-method-stat.txt](https://github.com/user-attachments/files/20905194/g1-method-stat.txt)
[g2-method-stat.txt](https://github.com/user-attachments/files/20906677/g2-method-stat.txt)

## Type of change

Please delete options that are not relevant.

- [x] New feature (non-breaking change which adds functionality)

# How has this been tested?

Scalar multiplication consistency has already been checked before using gopter.


# How has this been benchmarked?

G1
```
goos: linux
goarch: amd64
pkg: github.com/consensys/gnark-crypto/ecc/bls12-377
cpu: AMD Ryzen 9 7940HS w/ Radeon 780M Graphics     
        │      OLD      │                NEW                 │
        │    sec/op     │   sec/op     vs base               │
0         19387.0n ± 1%   655.2n ± 6%  -96.62% (p=0.002 n=6)
8          20.355µ ± 2%   2.468µ ± 1%  -87.88% (p=0.002 n=6)
16         21.205µ ± 1%   7.126µ ± 0%  -66.39% (p=0.002 n=6)
24         22.812µ ± 0%   9.316µ ± 1%  -59.16% (p=0.002 n=6)
32          23.74µ ± 1%   12.37µ ± 1%  -47.90% (p=0.002 n=6)
40          26.18µ ± 0%   15.80µ ± 0%  -39.65% (p=0.002 n=6)
48          26.61µ ± 1%   19.21µ ± 0%  -27.81% (p=0.002 n=6)
56          27.59µ ± 0%   21.81µ ± 0%  -20.93% (p=0.002 n=6)
64          29.25µ ± 0%   23.68µ ± 0%  -19.03% (p=0.002 n=6)
72          45.14µ ± 0%   26.73µ ± 0%  -40.78% (p=0.002 n=6)
80          45.96µ ± 1%   30.12µ ± 0%  -34.47% (p=0.002 n=6)
88          48.92µ ± 0%   34.11µ ± 0%  -30.27% (p=0.002 n=6)
96          49.90µ ± 1%   36.22µ ± 0%  -27.40% (p=0.002 n=6)
104         50.16µ ± 1%   37.63µ ± 0%  -24.98% (p=0.002 n=6)
112         49.37µ ± 1%   43.48µ ± 1%  -11.93% (p=0.002 n=6)
120         53.09µ ± 1%   43.64µ ± 0%  -17.81% (p=0.002 n=6)
128         52.25µ ± 1%   50.38µ ± 0%   -3.58% (p=0.002 n=6)
136         54.06µ ± 0%   51.46µ ± 1%   -4.81% (p=0.002 n=6)
144         53.32µ ± 0%   52.07µ ± 1%   -2.36% (p=0.002 n=6)
152         54.84µ ± 0%   54.78µ ± 0%        ~ (p=0.221 n=6)
160         52.51µ ± 1%   54.54µ ± 1%   +3.86% (p=0.002 n=6)
168         58.24µ ± 0%   53.75µ ± 0%   -7.72% (p=0.002 n=6)
176         54.13µ ± 0%   56.57µ ± 1%   +4.50% (p=0.002 n=6)
184         56.95µ ± 1%   57.85µ ± 1%   +1.58% (p=0.002 n=6)
192         55.35µ ± 1%   55.46µ ± 1%        ~ (p=0.240 n=6)
200         56.73µ ± 1%   56.29µ ± 0%   -0.78% (p=0.002 n=6)
208         57.00µ ± 0%   57.46µ ± 1%   +0.80% (p=0.002 n=6)
216         54.96µ ± 0%   57.13µ ± 0%   +3.96% (p=0.002 n=6)
224         56.27µ ± 1%   56.85µ ± 1%   +1.03% (p=0.004 n=6)
232         57.83µ ± 0%   56.64µ ± 1%   -2.05% (p=0.002 n=6)
240         59.08µ ± 0%   58.09µ ± 1%   -1.68% (p=0.002 n=6)
248         59.07µ ± 0%   58.00µ ± 0%   -1.81% (p=0.002 n=6)
geomean     42.69µ        29.57µ       -30.73%

pkg: github.com/consensys/gnark-crypto/ecc/bls12-381
        │      OLD      │                NEW                 │
        │    sec/op     │   sec/op     vs base               │
0         19823.5n ± 1%   661.1n ± 0%  -96.67% (p=0.002 n=6)
8          21.224µ ± 0%   3.317µ ± 0%  -84.37% (p=0.002 n=6)
16         21.217µ ± 0%   5.511µ ± 0%  -74.03% (p=0.002 n=6)
24         23.590µ ± 1%   9.374µ ± 0%  -60.26% (p=0.002 n=6)
32          24.02µ ± 0%   12.80µ ± 1%  -46.74% (p=0.002 n=6)
40          25.27µ ± 1%   14.66µ ± 1%  -41.97% (p=0.002 n=6)
48          26.05µ ± 1%   17.62µ ± 1%  -32.36% (p=0.002 n=6)
56          28.53µ ± 1%   21.45µ ± 0%  -24.83% (p=0.002 n=6)
64          28.47µ ± 0%   23.25µ ± 0%  -18.35% (p=0.002 n=6)
72          46.41µ ± 0%   27.93µ ± 0%  -39.82% (p=0.002 n=6)
80          46.74µ ± 1%   30.50µ ± 1%  -34.73% (p=0.002 n=6)
88          48.00µ ± 0%   33.15µ ± 0%  -30.93% (p=0.002 n=6)
96          49.75µ ± 1%   38.62µ ± 0%  -22.36% (p=0.002 n=6)
104         48.41µ ± 0%   40.01µ ± 0%  -17.35% (p=0.002 n=6)
112         50.93µ ± 0%   41.40µ ± 1%  -18.73% (p=0.002 n=6)
120         51.25µ ± 0%   46.90µ ± 0%   -8.48% (p=0.002 n=6)
128         55.76µ ± 0%   50.42µ ± 1%   -9.58% (p=0.002 n=6)
136         54.83µ ± 1%   53.75µ ± 0%   -1.97% (p=0.002 n=6)
144         55.69µ ± 2%   54.32µ ± 0%   -2.46% (p=0.002 n=6)
152         54.58µ ± 0%   56.33µ ± 0%   +3.22% (p=0.002 n=6)
160         54.16µ ± 0%   55.38µ ± 0%   +2.24% (p=0.002 n=6)
168         56.22µ ± 0%   55.82µ ± 0%   -0.72% (p=0.002 n=6)
176         56.27µ ± 0%   55.02µ ± 1%   -2.22% (p=0.002 n=6)
184         54.87µ ± 0%   56.17µ ± 0%   +2.38% (p=0.002 n=6)
192         57.19µ ± 0%   55.88µ ± 0%   -2.28% (p=0.002 n=6)
200         56.52µ ± 0%   56.43µ ± 0%        ~ (p=0.240 n=6)
208         57.59µ ± 0%   58.46µ ± 0%   +1.51% (p=0.002 n=6)
216         57.09µ ± 0%   57.55µ ± 0%   +0.81% (p=0.002 n=6)
224         57.58µ ± 0%   57.58µ ± 1%        ~ (p=0.554 n=6)
232         57.17µ ± 1%   59.21µ ± 0%   +3.57% (p=0.002 n=6)
240         58.45µ ± 0%   58.39µ ± 0%        ~ (p=0.180 n=6)
248         58.10µ ± 0%   58.78µ ± 1%   +1.18% (p=0.002 n=6)
geomean     42.98µ        29.84µ       -30.58%

pkg: github.com/consensys/gnark-crypto/ecc/bls24-315
        │      OLD      │                NEW                 │
        │    sec/op     │   sec/op     vs base               │
0         15535.0n ± 1%   521.4n ± 0%  -96.64% (p=0.002 n=6)
8          16.767µ ± 0%   2.931µ ± 0%  -82.52% (p=0.002 n=6)
16         17.058µ ± 0%   4.657µ ± 1%  -72.70% (p=0.002 n=6)
24         18.636µ ± 1%   7.044µ ± 0%  -62.20% (p=0.002 n=6)
32          19.55µ ± 1%   10.71µ ± 0%  -45.23% (p=0.002 n=6)
40          18.96µ ± 0%   12.44µ ± 0%  -34.41% (p=0.002 n=6)
48          21.48µ ± 0%   14.14µ ± 0%  -34.16% (p=0.002 n=6)
56          20.51µ ± 1%   16.79µ ± 0%  -18.13% (p=0.002 n=6)
64          23.31µ ± 0%   18.25µ ± 0%  -21.71% (p=0.002 n=6)
72          36.31µ ± 1%   21.91µ ± 1%  -39.66% (p=0.002 n=6)
80          37.12µ ± 0%   23.37µ ± 0%  -37.04% (p=0.002 n=6)
88          38.19µ ± 1%   25.37µ ± 0%  -33.56% (p=0.002 n=6)
96          36.85µ ± 0%   28.34µ ± 0%  -23.10% (p=0.002 n=6)
104         40.04µ ± 0%   31.38µ ± 1%  -21.65% (p=0.002 n=6)
112         39.91µ ± 1%   33.38µ ± 0%  -16.35% (p=0.002 n=6)
120         40.27µ ± 0%   36.76µ ± 0%   -8.73% (p=0.002 n=6)
128         42.80µ ± 0%   35.33µ ± 1%  -17.47% (p=0.002 n=6)
136         42.55µ ± 1%   41.52µ ± 0%   -2.40% (p=0.002 n=6)
144         42.15µ ± 0%   40.33µ ± 0%   -4.31% (p=0.002 n=6)
152         42.21µ ± 0%   44.19µ ± 0%   +4.68% (p=0.002 n=6)
160         41.81µ ± 0%   42.28µ ± 0%   +1.11% (p=0.002 n=6)
168         42.16µ ± 0%   43.51µ ± 0%   +3.20% (p=0.002 n=6)
176         42.47µ ± 0%   43.89µ ± 0%   +3.33% (p=0.002 n=6)
184         44.51µ ± 1%   43.21µ ± 0%   -2.91% (p=0.002 n=6)
192         43.90µ ± 0%   44.85µ ± 0%   +2.15% (p=0.002 n=6)
200         44.78µ ± 1%   44.15µ ± 0%   -1.39% (p=0.002 n=6)
208         44.29µ ± 0%   44.47µ ± 0%   +0.41% (p=0.009 n=6)
216         45.49µ ± 0%   43.24µ ± 0%   -4.93% (p=0.002 n=6)
224         44.49µ ± 0%   44.96µ ± 1%   +1.06% (p=0.002 n=6)
232         45.23µ ± 1%   44.88µ ± 0%   -0.77% (p=0.002 n=6)
240         45.09µ ± 0%   46.20µ ± 0%   +2.45% (p=0.002 n=6)
248         46.37µ ± 0%   45.79µ ± 0%   -1.24% (p=0.002 n=6)
geomean     33.59µ        23.35µ       -30.50%

pkg: github.com/consensys/gnark-crypto/ecc/bls24-317
        │      OLD      │                NEW                 │
        │    sec/op     │   sec/op     vs base               │
0         15299.0n ± 0%   521.9n ± 0%  -96.59% (p=0.002 n=6)
8          16.090µ ± 0%   1.951µ ± 0%  -87.87% (p=0.002 n=6)
16         17.067µ ± 0%   4.648µ ± 0%  -72.76% (p=0.002 n=6)
24         17.722µ ± 0%   6.708µ ± 1%  -62.15% (p=0.002 n=6)
32         18.429µ ± 1%   9.699µ ± 0%  -47.37% (p=0.002 n=6)
40          19.36µ ± 0%   11.12µ ± 0%  -42.57% (p=0.002 n=6)
48          20.67µ ± 0%   13.20µ ± 0%  -36.15% (p=0.002 n=6)
56          21.35µ ± 0%   16.83µ ± 1%  -21.16% (p=0.002 n=6)
64          22.69µ ± 0%   17.34µ ± 1%  -23.56% (p=0.002 n=6)
72          35.63µ ± 0%   21.32µ ± 0%  -40.17% (p=0.002 n=6)
80          35.93µ ± 0%   25.56µ ± 1%  -28.87% (p=0.002 n=6)
88          39.26µ ± 1%   26.97µ ± 0%  -31.31% (p=0.002 n=6)
96          37.56µ ± 0%   28.07µ ± 0%  -25.25% (p=0.002 n=6)
104         38.91µ ± 0%   29.52µ ± 1%  -24.13% (p=0.002 n=6)
112         40.92µ ± 0%   34.78µ ± 1%  -15.00% (p=0.002 n=6)
120         42.80µ ± 0%   36.51µ ± 0%  -14.69% (p=0.002 n=6)
128         43.44µ ± 0%   38.52µ ± 0%  -11.33% (p=0.002 n=6)
136         41.35µ ± 1%   39.94µ ± 1%   -3.40% (p=0.002 n=6)
144         41.57µ ± 1%   41.77µ ± 0%   +0.47% (p=0.041 n=6)
152         41.20µ ± 0%   42.00µ ± 0%   +1.95% (p=0.002 n=6)
160         41.94µ ± 0%   42.88µ ± 0%   +2.24% (p=0.002 n=6)
168         43.47µ ± 0%   41.64µ ± 0%   -4.22% (p=0.002 n=6)
176         44.10µ ± 0%   44.29µ ± 0%   +0.44% (p=0.002 n=6)
184         44.51µ ± 1%   43.62µ ± 1%   -2.00% (p=0.002 n=6)
192         44.05µ ± 0%   44.60µ ± 1%   +1.25% (p=0.002 n=6)
200         42.90µ ± 0%   47.13µ ± 0%   +9.86% (p=0.002 n=6)
208         44.10µ ± 0%   44.49µ ± 0%   +0.87% (p=0.002 n=6)
216         44.45µ ± 0%   46.13µ ± 0%   +3.79% (p=0.002 n=6)
224         45.73µ ± 0%   44.90µ ± 0%   -1.81% (p=0.002 n=6)
232         45.06µ ± 1%   45.88µ ± 0%   +1.83% (p=0.002 n=6)
240         44.08µ ± 1%   46.82µ ± 0%   +6.23% (p=0.002 n=6)
248         46.47µ ± 0%   46.16µ ± 0%   -0.67% (p=0.002 n=6)
geomean     33.42µ        22.98µ       -31.25%

pkg: github.com/consensys/gnark-crypto/ecc/bn254
        │      OLD      │                NEW                 │
        │    sec/op     │   sec/op     vs base               │
0         10963.0n ± 0%   365.1n ± 1%  -96.67% (p=0.002 n=6)
8          11.593µ ± 0%   1.385µ ± 0%  -88.05% (p=0.002 n=6)
16         11.813µ ± 0%   3.337µ ± 1%  -71.75% (p=0.002 n=6)
24         13.206µ ± 0%   5.264µ ± 0%  -60.14% (p=0.002 n=6)
32         13.710µ ± 2%   6.480µ ± 1%  -52.73% (p=0.002 n=6)
40         14.130µ ± 0%   8.899µ ± 0%  -37.02% (p=0.002 n=6)
48          15.79µ ± 0%   10.57µ ± 0%  -33.09% (p=0.002 n=6)
56          15.80µ ± 0%   11.15µ ± 0%  -29.42% (p=0.002 n=6)
64          16.52µ ± 0%   14.02µ ± 0%  -15.11% (p=0.002 n=6)
72          24.90µ ± 0%   14.72µ ± 0%  -40.86% (p=0.002 n=6)
80          26.69µ ± 0%   16.92µ ± 1%  -36.60% (p=0.002 n=6)
88          27.41µ ± 2%   17.93µ ± 0%  -34.58% (p=0.002 n=6)
96          28.74µ ± 0%   21.05µ ± 1%  -26.76% (p=0.002 n=6)
104         27.15µ ± 1%   22.21µ ± 1%  -18.20% (p=0.002 n=6)
112         29.64µ ± 0%   23.07µ ± 0%  -22.18% (p=0.002 n=6)
120         30.09µ ± 0%   23.77µ ± 0%  -21.02% (p=0.002 n=6)
128         30.86µ ± 0%   27.17µ ± 0%  -11.97% (p=0.002 n=6)
136         31.22µ ± 0%   29.96µ ± 0%   -4.05% (p=0.002 n=6)
144         32.35µ ± 2%   32.48µ ± 1%        ~ (p=0.180 n=6)
152         33.04µ ± 0%   32.89µ ± 2%        ~ (p=0.065 n=6)
160         31.50µ ± 0%   32.87µ ± 0%   +4.33% (p=0.002 n=6)
168         32.55µ ± 0%   32.86µ ± 0%   +0.97% (p=0.002 n=6)
176         32.98µ ± 0%   32.83µ ± 0%   -0.46% (p=0.002 n=6)
184         31.96µ ± 0%   32.65µ ± 1%   +2.15% (p=0.002 n=6)
192         32.37µ ± 0%   32.60µ ± 0%   +0.72% (p=0.002 n=6)
200         33.09µ ± 1%   33.71µ ± 1%   +1.89% (p=0.002 n=6)
208         32.17µ ± 0%   32.51µ ± 0%   +1.05% (p=0.002 n=6)
216         33.45µ ± 0%   32.93µ ± 0%   -1.55% (p=0.002 n=6)
224         33.03µ ± 0%   32.93µ ± 1%        ~ (p=0.093 n=6)
232         33.23µ ± 0%   33.37µ ± 0%   +0.41% (p=0.009 n=6)
240         33.71µ ± 0%   33.72µ ± 0%        ~ (p=0.974 n=6)
248         33.27µ ± 1%   33.54µ ± 0%   +0.80% (p=0.026 n=6)
geomean     24.58µ        16.73µ       -31.94%

pkg: github.com/consensys/gnark-crypto/ecc/bw6-633
        │     OLD      │                NEW                 │
        │    sec/op    │   sec/op     vs base               │
0         49.474µ ± 0%   1.660µ ± 1%  -96.64% (p=0.002 n=6)
8         52.756µ ± 1%   9.395µ ± 0%  -82.19% (p=0.002 n=6)
16         53.86µ ± 1%   17.03µ ± 0%  -68.38% (p=0.002 n=6)
24         56.48µ ± 2%   24.73µ ± 0%  -56.21% (p=0.002 n=6)
32         64.59µ ± 0%   29.28µ ± 0%  -54.67% (p=0.002 n=6)
40         63.71µ ± 1%   37.85µ ± 0%  -40.60% (p=0.002 n=6)
48         65.64µ ± 0%   45.69µ ± 0%  -30.40% (p=0.002 n=6)
56         70.10µ ± 0%   54.39µ ± 1%  -22.42% (p=0.002 n=6)
64         72.93µ ± 1%   61.99µ ± 0%  -15.00% (p=0.002 n=6)
72        113.86µ ± 0%   65.46µ ± 0%  -42.51% (p=0.002 n=6)
80        118.33µ ± 0%   75.16µ ± 0%  -36.48% (p=0.002 n=6)
88        121.59µ ± 1%   86.88µ ± 0%  -28.54% (p=0.002 n=6)
96        120.94µ ± 0%   93.72µ ± 1%  -22.51% (p=0.002 n=6)
104        126.9µ ± 0%   101.7µ ± 0%  -19.87% (p=0.002 n=6)
112        132.9µ ± 1%   105.1µ ± 0%  -20.92% (p=0.002 n=6)
120        134.1µ ± 1%   110.5µ ± 1%  -17.54% (p=0.002 n=6)
128        137.4µ ± 0%   123.5µ ± 0%  -10.08% (p=0.002 n=6)
136        177.4µ ± 0%   133.4µ ± 0%  -24.82% (p=0.002 n=6)
144        180.4µ ± 1%   141.2µ ± 0%  -21.76% (p=0.002 n=6)
152        179.7µ ± 0%   145.8µ ± 0%  -18.86% (p=0.002 n=6)
160        188.4µ ± 0%   153.6µ ± 1%  -18.49% (p=0.002 n=6)
168        191.1µ ± 0%   164.2µ ± 0%  -14.09% (p=0.002 n=6)
176        187.9µ ± 0%   163.2µ ± 0%  -13.12% (p=0.002 n=6)
184        185.7µ ± 0%   185.6µ ± 0%        ~ (p=0.240 n=6)
192        189.9µ ± 1%   191.6µ ± 1%   +0.93% (p=0.002 n=6)
200        193.4µ ± 1%   190.8µ ± 0%   -1.37% (p=0.002 n=6)
208        190.9µ ± 0%   188.8µ ± 0%   -1.08% (p=0.002 n=6)
216        195.3µ ± 0%   196.1µ ± 0%   +0.41% (p=0.004 n=6)
224        193.3µ ± 1%   196.2µ ± 1%   +1.52% (p=0.002 n=6)
232        193.4µ ± 0%   196.7µ ± 1%   +1.68% (p=0.002 n=6)
240        198.6µ ± 0%   201.5µ ± 1%   +1.46% (p=0.002 n=6)
248        198.8µ ± 0%   198.2µ ± 0%   -0.31% (p=0.015 n=6)
256        198.8µ ± 1%   197.0µ ± 0%   -0.92% (p=0.002 n=6)
264        196.7µ ± 0%   196.1µ ± 0%        ~ (p=0.394 n=6)
272        198.6µ ± 0%   199.3µ ± 0%   +0.37% (p=0.002 n=6)
280        198.7µ ± 0%   201.5µ ± 0%   +1.40% (p=0.002 n=6)
288        198.9µ ± 0%   200.4µ ± 1%   +0.72% (p=0.002 n=6)
296        201.2µ ± 0%   202.5µ ± 1%   +0.66% (p=0.002 n=6)
304        199.5µ ± 1%   196.1µ ± 0%   -1.67% (p=0.002 n=6)
312        198.6µ ± 0%   201.6µ ± 0%   +1.52% (p=0.002 n=6)
geomean    136.3µ        99.16µ       -27.26%

pkg: github.com/consensys/gnark-crypto/ecc/bw6-761
        │     OLD      │                NEW                 │
        │    sec/op    │   sec/op     vs base               │
0         63.790µ ± 0%   2.162µ ± 0%  -96.61% (p=0.002 n=6)
8          66.82µ ± 0%   12.19µ ± 0%  -81.75% (p=0.002 n=6)
16         69.52µ ± 0%   20.80µ ± 0%  -70.08% (p=0.002 n=6)
24         76.54µ ± 1%   32.22µ ± 0%  -57.90% (p=0.002 n=6)
32         82.16µ ± 1%   42.29µ ± 1%  -48.53% (p=0.002 n=6)
40         79.45µ ± 1%   50.94µ ± 1%  -35.88% (p=0.002 n=6)
48         86.27µ ± 0%   63.61µ ± 0%  -26.27% (p=0.002 n=6)
56         90.40µ ± 1%   74.97µ ± 0%  -17.07% (p=0.002 n=6)
64         97.42µ ± 1%   73.82µ ± 0%  -24.22% (p=0.002 n=6)
72        147.18µ ± 0%   92.30µ ± 0%  -37.28% (p=0.002 n=6)
80        149.95µ ± 1%   95.37µ ± 1%  -36.40% (p=0.002 n=6)
88         154.5µ ± 0%   102.8µ ± 1%  -33.47% (p=0.002 n=6)
96         165.4µ ± 0%   122.4µ ± 1%  -26.01% (p=0.002 n=6)
104        170.9µ ± 0%   126.8µ ± 0%  -25.79% (p=0.002 n=6)
112        155.9µ ± 0%   135.6µ ± 0%  -13.03% (p=0.002 n=6)
120        171.3µ ± 0%   153.8µ ± 0%  -10.26% (p=0.002 n=6)
128        178.0µ ± 1%   170.9µ ± 0%   -3.98% (p=0.002 n=6)
136        224.9µ ± 0%   175.3µ ± 0%  -22.06% (p=0.002 n=6)
144        234.3µ ± 0%   182.2µ ± 1%  -22.24% (p=0.002 n=6)
152        230.2µ ± 0%   191.3µ ± 0%  -16.91% (p=0.002 n=6)
160        235.8µ ± 0%   208.2µ ± 0%  -11.68% (p=0.002 n=6)
168        241.3µ ± 0%   211.3µ ± 0%  -12.40% (p=0.002 n=6)
176        245.5µ ± 0%   222.4µ ± 0%   -9.39% (p=0.002 n=6)
184        255.4µ ± 0%   241.0µ ± 0%   -5.62% (p=0.002 n=6)
192        259.7µ ± 1%   241.2µ ± 1%   -7.12% (p=0.002 n=6)
200        256.9µ ± 0%   253.6µ ± 0%   -1.31% (p=0.002 n=6)
208        262.1µ ± 0%   260.6µ ± 1%        ~ (p=0.065 n=6)
216        263.6µ ± 0%   265.9µ ± 0%   +0.87% (p=0.002 n=6)
224        273.0µ ± 0%   269.8µ ± 0%   -1.17% (p=0.002 n=6)
232        262.2µ ± 0%   267.5µ ± 0%   +2.00% (p=0.002 n=6)
240        270.6µ ± 0%   271.5µ ± 0%   +0.32% (p=0.002 n=6)
248        262.2µ ± 0%   271.7µ ± 0%   +3.63% (p=0.002 n=6)
256        274.5µ ± 1%   279.8µ ± 1%   +1.93% (p=0.002 n=6)
264        273.2µ ± 0%   268.6µ ± 0%   -1.68% (p=0.002 n=6)
272        270.6µ ± 0%   270.7µ ± 1%        ~ (p=1.000 n=6)
280        272.2µ ± 0%   269.4µ ± 1%   -1.02% (p=0.002 n=6)
288        276.0µ ± 0%   274.2µ ± 0%   -0.68% (p=0.002 n=6)
296        278.8µ ± 0%   277.0µ ± 0%   -0.63% (p=0.002 n=6)
304        279.9µ ± 0%   287.0µ ± 1%   +2.53% (p=0.002 n=6)
312        281.8µ ± 0%   280.0µ ± 0%   -0.61% (p=0.002 n=6)
320        276.4µ ± 1%   275.8µ ± 0%   -0.20% (p=0.009 n=6)
328        279.0µ ± 0%   282.5µ ± 0%   +1.24% (p=0.002 n=6)
336        280.4µ ± 0%   277.1µ ± 0%   -1.15% (p=0.002 n=6)
344        276.0µ ± 0%   284.2µ ± 2%   +2.95% (p=0.002 n=6)
352        270.8µ ± 0%   277.2µ ± 0%   +2.37% (p=0.002 n=6)
360        281.5µ ± 0%   285.3µ ± 1%   +1.35% (p=0.002 n=6)
368        278.8µ ± 0%   281.4µ ± 0%   +0.91% (p=0.002 n=6)
376        273.3µ ± 0%   281.2µ ± 0%   +2.89% (p=0.002 n=6)
geomean    193.7µ        150.0µ       -22.58%

pkg: github.com/consensys/gnark-crypto/ecc/grumpkin
        │      OLD      │                NEW                 │
        │    sec/op     │   sec/op     vs base               │
0         10976.0n ± 0%   368.1n ± 0%  -96.65% (p=0.002 n=6)
8          11.798µ ± 1%   2.088µ ± 0%  -82.30% (p=0.002 n=6)
16         12.285µ ± 0%   3.570µ ± 0%  -70.94% (p=0.002 n=6)
24         12.944µ ± 0%   4.798µ ± 1%  -62.93% (p=0.002 n=6)
32         13.457µ ± 0%   6.077µ ± 0%  -54.84% (p=0.002 n=6)
40         13.912µ ± 0%   8.425µ ± 0%  -39.44% (p=0.002 n=6)
48          15.07µ ± 0%   10.82µ ± 0%  -28.19% (p=0.002 n=6)
56          15.59µ ± 0%   11.61µ ± 0%  -25.55% (p=0.002 n=6)
64          17.64µ ± 1%   14.27µ ± 0%  -19.13% (p=0.002 n=6)
72          25.30µ ± 0%   16.16µ ± 0%  -36.11% (p=0.002 n=6)
80          25.84µ ± 0%   16.94µ ± 0%  -34.42% (p=0.002 n=6)
88          26.35µ ± 0%   18.88µ ± 0%  -28.35% (p=0.002 n=6)
96          26.77µ ± 0%   19.76µ ± 0%  -26.18% (p=0.002 n=6)
104         27.94µ ± 0%   21.85µ ± 0%  -21.79% (p=0.002 n=6)
112         29.29µ ± 0%   24.50µ ± 0%  -16.37% (p=0.002 n=6)
120         28.61µ ± 1%   24.80µ ± 0%  -13.32% (p=0.002 n=6)
128         32.18µ ± 0%   26.48µ ± 0%  -17.72% (p=0.002 n=6)
136         31.05µ ± 0%   29.33µ ± 1%   -5.53% (p=0.002 n=6)
144         30.89µ ± 0%   32.71µ ± 0%   +5.91% (p=0.002 n=6)
152         32.60µ ± 1%   32.32µ ± 0%   -0.85% (p=0.002 n=6)
160         32.48µ ± 0%   32.98µ ± 0%   +1.54% (p=0.002 n=6)
168         32.50µ ± 0%   32.74µ ± 1%   +0.73% (p=0.002 n=6)
176         32.45µ ± 1%   32.70µ ± 0%   +0.75% (p=0.041 n=6)
184         33.06µ ± 1%   32.05µ ± 0%   -3.06% (p=0.002 n=6)
192         33.28µ ± 2%   32.36µ ± 1%   -2.78% (p=0.002 n=6)
200         32.86µ ± 0%   33.23µ ± 0%   +1.12% (p=0.002 n=6)
208         33.71µ ± 0%   33.46µ ± 0%   -0.75% (p=0.002 n=6)
216         33.50µ ± 0%   33.62µ ± 0%   +0.34% (p=0.026 n=6)
224         33.89µ ± 1%   32.80µ ± 0%   -3.22% (p=0.002 n=6)
232         32.30µ ± 0%   33.22µ ± 0%   +2.84% (p=0.002 n=6)
240         33.52µ ± 0%   33.13µ ± 1%   -1.16% (p=0.004 n=6)
248         32.59µ ± 1%   33.30µ ± 0%   +2.18% (p=0.002 n=6)
geomean     24.52µ        16.96µ       -30.81%

pkg: github.com/consensys/gnark-crypto/ecc/secp256k1
        │      OLD      │                NEW                 │
…