@ivokub 2025-06-25
Converted to draft - G2 benchmarks are incorrect (thanks copilot)